+++ src/graphql_clj/box.clj
@@ -1,6 +1,5 @@
 (ns graphql-clj.box
-  (:import [clojure.lang IObj IDeref IPersistentCollection Named]
-           [java.io Writer]))
+  (:import [clojure.lang IObj IDeref IPersistentCollection Named]))
 
 (deftype Box [value meta]
 
@@ -20,9 +19,6 @@
   (equals [_ b] (if (instance? Box b) (= value @b) (= value b)))
   (hashCode [_] (.hashCode value))
   (toString [_] (str value)))
-
-(defmethod print-method Box [v ^Writer w]
-  (.write w (.toString v)))
 
 (defn kv->box
   "Given a single key value pair in a map, preserve metadata for the value."