+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file. This change
 - Fix bug in validation of unused fragments
 - Fix bugs in validation of scalar leafs for lists of scalars and lists of objects
 - Deprecate `graphql-clj.type/create-schema` in favor of `graphql-clj.validator/validate-schema`
+- Fix bug in fragment cycles validation
+- Properly assign specs for recursively nested fields (switch to pre-order traversal for adding statement specs)
 
 ## [0.1.18] - 2016-10-20
 - Validate that statements do not contain fragment cycles